Add skip_invalid_emails feature and upgrade default behavior
There are two distinct KeyError cases that needed to be handled. The first arises when an old email address is not found or is for a deactivated account. The second arises when a desired new email address is already in use.

The script will have an optional command line argument called "--skip_invalid_emails" that can enable the feature to bypass these KeyErrors and continue looping through the list of member_emails to update.

In order to accommodate the "skip_invalid_emails" feature and to upgrade the default behavior, I inserted try/except blocks so that aborting or continuing the script could be controlled at the points at which the two cases of KeyError could occur. As a side-effect of this, each attempted email update will be POSTed individually rather than combined into a single payload. As a result of this, I could also improve the script's logging so that each update attempt for each email will have its success/fail outcome explicitly printed to the terminal with a reason stated for why if it failed.

Lastly, I improved the default behavior so that when KeyError is excepted the script will abort gracefully with exit code 1.