Gary/extend batch update users email (#9)
* Make get_access_token response print explicit

* Add optional arg for enabling skip invalid emails feature

* Add skip_invalid_emails feature and upgrade default behavior

There are two distinct KeyError cases that needed to be handled. The first arises when an old email address is not found or is for a deactivated account. The second arises when a desired new email address is already in use.

The script will have an optional command line argument called "--skip_invalid_emails" that can enable the feature to bypass these KeyErrors and continue looping through the list of member_emails to update.

In order to accommodate the "skip_invalid_emails" feature and to upgrade the default behavior, I inserted try/except blocks so that aborting or continuing the script could be controlled at the points at which the two cases of KeyError could occur. As a side-effect of this, each attempted email update will be POSTed individually rather than combined into a single payload. As a result of this, I could also improve the script's logging so that each update attempt for each email will have its success/fail outcome explicitly printed to the terminal with a reason stated for why if it failed.

Lastly, I improved the default behavior so that when KeyError is excepted the script will abort gracefully with exit code 1.

* Add extra valid headers and better example rows

* Improve docs

- Update instructions to be relevant to current state of the script.
- Include more useful links and explainer notes

* Extend and upgrade script

- Functionality: Add ability to update additional user attributes: First Name, Last Name, Member Type
- Functionality: Add option to abort script when an attempted update encounters an error rather than continuing to update next member
- Guardrails: In all instances where requests to API are made, use try/except blocks to raise exceptions for errors and surface info to user via handling and print statements in downstream code
- Guardrails: Check that required Email header column is present in CSV, notify and abort if not
- UX: improve logging print statements, including: what the script is doing while it is running, whether updates succeeded or errored, what values users' attributes have been updated to, specifics about why an error occurred and how to potentially solve it, when the script is aborted

* In cloud arg help, specify where to check in Sigma

* Replace "metadata" with term we use in our docs: "user attributes"